# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API для управления студентами, классами и лекциями",
        "version": "1.0"
    },
    "paths": {
...
        "/api/classes": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Получить список всех классов",
                "operationId": "1ea478be0955b21841689b73bb18c03b",
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/StudentClass"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Classes"
                ],
                "summary": "Создать класс",
                "operationId": "7deb499f28967400f4e21c7993d30d72",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "10-A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Класс создан",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentClass"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
        "/api/classes/{id}": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Получить информацию о конкретном классе",
                "operationId": "7a0f9a169189cf59ccec56d59d643abc",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Успешный ответ",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentClass"
                                        },
                                        "students": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Student"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Класс не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Classes"
                ],
                "summary": "Обновить класс",
                "operationId": "884647da0568faf7dc09d48b272d4f19",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "10-B"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Класс обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StudentClass"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Класс не найден"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Classes"
                ],
                "summary": "Удалить класс",
                "operationId": "70cb1e83754f4de1b8dcab6a77699b3d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Класс удален"
                    },
                    "404": {
                        "description": "Класс не найден"
                    }
                }
            }
        },
        "/api/classes/{id}/curriculum": {
            "get": {
                "tags": [
                    "Classes"
                ],
                "summary": "Получить учебный план класса",
                "operationId": "c3f79adba20565885f7e68716fc9eeb7",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Учебный план",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/StudentClass"
                                        },
                                        "lectures": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer"
                                                    },
                                                    "topic": {
                                                        "type": "string"
                                                    },
                                                    "description": {
                                                        "type": "string"
                                                    },
                                                    "order": {
                                                        "type": "integer"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Класс не найден"
                    }
                }
            },
            "put": {
                "tags": [
                    "Classes"
                ],
                "summary": "Обновить учебный план класса",
                "operationId": "6eb76ccd3a0752961d14e4de79a0908c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateClassCurriculumRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Учебный план обновлен",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer"
                                            },
                                            "topic": {
                                                "type": "string"
                                            },
                                            "description": {
                                                "type": "string"
                                            },
                                            "order": {
                                                "type": "integer"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Класс не найден"
                    },
                    "422": {
                        "description": "Ошибка валидации"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UpdateClassCurriculumRequest": {
                "required": [
                    "lectures"
                ],
                "properties": {
                    "lectures": {
                        "type": "array",
                        "items": {
                            "required": [
                                "lecture_id",
                                "order"
                            ],
                            "properties": {
                                "lecture_id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "order": {
                                    "type": "integer",
                                    "example": 0
                                }
                            },
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
            "Student": {
                "title": "Student",
                "description": "Модель студента",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Иван Иванов"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "ivan@example.com"
                    },
                    "student_class_id": {
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "StudentClass": {
                "title": "StudentClass",
                "description": "Модель класса",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "10-A"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            }
        }
    },
...
}